move authFlowType to cognito module
